+++ application/controllers/Pencegahan.php
@@ -123,6 +123,10 @@ class Pencegahan extends CI_Controller
         }
     }
     public function delete_data($id){
+        if($this->session->userdata('peringkat') != "admin"){
+			redirect(base_url("/dashboard"));
+		}
+
         $validasi = $this->Pencegahan_model->get_where($id);
         if (!empty($validasi)) {
             $this->Pencegahan_model->delete($id);
