+++ application/controllers/Sdm.php
@@ -122,6 +122,9 @@ class Sdm extends CI_Controller
         }
     }
     public function delete_data($id){
+        if($this->session->userdata('peringkat') != "admin"){
+			redirect(base_url("/dashboard"));
+		}
         $validasi = $this->Sdm_model->get_where($id);
         if (!empty($validasi)) {
             $this->Sdm_model->delete($id);
